gestion carburant 1

+++ app/Database/Migrations/2023-08-02-163618_Ravitaillements.php
@@ -33,7 +33,7 @@ class Ravitaillements extends Migration
                 'null' => true,
             ],
             'prix_litre' => [
-                'type' => 'INT',
+                'type' => 'FLOAT',
                 'constraint' => 11,
                 'default' => 0,
                 'null' => true,
